feat!: add more algorithms
The program now has 3 more other scientific approaches to calculate a readability score besides the ARI. It calculates the score and the age accordingly.

The first algorithm is Flesch–Kincaid readability tests.

The second one is SMOG index. It stands for Simple Measure of Gobbledygook.

The next one is Coleman–Liau index.