feat: start the Battleship project
Create base classes for the game. The Main class to start the game, the Printer class to print elements to the console. The Symbols enum to set symbols representing the board. The Board class to organize and record the battlefield. And the Game class with all its logic and data processing.